% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add 2 numerical parameters}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{the first number to add.}

\item{y}{the second number to add which is 10 by default, unless mentioned.}
}
\value{
A number depicting the sum of x and y.
}
\description{
Add 2 numerical parameters
}
\examples{
my_add(1, 20)
my_add(1)

}
